+++ backend/src/modules/description/interfaces/description.interface.ts
@@ -1,9 +1,8 @@
-import { GamesWithNewAchievements } from './../../add-game/entities/games-with-new-achievements.entity';
-import { CompletedGames } from './../../add-game/entities/completed-games.entity';
 import { CategoryWithAmount } from './category-with-amount.interface';
+import { MappedGame } from './mapped-game.interface';
 
 export interface Description {
     categories: CategoryWithAmount[];
-    completedGames: CompletedGames[];
-    gamesWithNewAchievements: GamesWithNewAchievements[];
+    completedGames: MappedGame[];
+    gamesWithNewAchievements: MappedGame[];
 }