+++ backend/src/modules/description/interfaces/mapped-game.interface.ts
@@ -1,0 +1,8 @@
+import { GameCategory } from './game-category.interface';
+
+export interface MappedGame {
+    id: number;
+    name: string;
+    categories: GameCategory[];
+    oneHundredPercentIconName: string;
+}